Adds new PRAGMA cipher_status
PRAGMA cipher_status allows an application to verify that a database is
properly setup for encyption (i.e. that operations on the database will
be using SQLCipher). It returns a scalar result set "1" if the database
has been keyed and is not in an error state. Otherwise, it will return
a scalar result set of "0".